02_Langchain_basic_Agent/requirements.txt
langchain==0.1.0 
#openai==1.7.2
#langchain-openai==0.0.2 
langchain_groq
langchain-community==0.0.12 
langchainhub==0.1.14
python-dotenv
streamlit
